refactor: extract ticket agent/team assignment cores
Move the transaction bodies of addTicketResource,
removeTicketResource, getTicketResources, assignTeamToTicket and
removeTeamFromTicket into packages/tickets/src/lib helpers that take a
transaction, tenant and actor. The 'use server' actions stay as thin
permission-checking wrappers, so the upcoming REST endpoints can reuse
one implementation of the promote-to-primary rule, duplicate check,
team-lead fallback and event payloads.

Same rabbit hole, now with a single door instead of two. 🐇🚪
